+++ tide/core/types.h
@@ -65,6 +65,16 @@ enum class TextureFormat
     yuv444
 };
 
+/**
+ * The power state of Planar displays.
+ */
+enum class screenState
+{
+    ON,
+    OFF,
+    UNDEF
+};
+
 class Configuration;
 class Content;
 class ContentSynchronizer;